Add dashboard socket connection
