+++ package.json
@@ -14,7 +14,8 @@
   "author": "Rick van der Staaij",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.13.4"
+    "express": "~4.13.4",
+    "socket.io": "~1.4.5"
   },
   "devDependencies": {
     "gulp-develop-server": "~0.5.1",
