+++ public/index.html
@@ -1,0 +1,15 @@
+<html>
+<head>
+    <title>CI Montior</title>
+</head>
+<body>
+    <script src="/socket.io/socket.io.js"></script>
+    <script>
+        var socket = io();
+    </script>
+    <h1>CI Monitor</h1>
+    <div id="list">
+
+    </div>
+</body>
+</html>
